+++ supabase/migrations/20251203000000_fix_beach_affinity_trigger_security.sql
@@ -57,3 +57,6 @@ COMMENT ON FUNCTION update_beach_affinity_on_session_change IS 'Trigger function
 
 
 
+
+
+